Update regression test number 289, which now runs successfully.
This test works as of revision db08bfafa9a67e20fc3ddc5157075cc471ea2b04.

Versions of SBV prior to 3.1 did not distinguish the boolean type from
the size-1 bitvector type; this example uses literals like "0b0" and
"0b1", so it triggered the SBV bug, which manifested as a type error from
the external SMT solver.