Change the representation of type substitutions.
Now, substitions are explicitly a sequence of maps from variables
to types.  This means that composing substitions is nearly trivial
because the structure of maps does not need to be rebuilt.  Instead,
we traverse the sequence of maps when substituions are applied.

This significantly reduces the garbage produced by type substituion.