Canonicalize module path names earlier
This largely reverts 54831eace2801c872261b86a3e7d19f19656e134 and takes
a different approach. This way path names are normalized before feeding
them to the parser, rather than when comparing the module to be loaded
with the module already loaded. This works better because it takes into
account the current directory at the time when `loadModuleByPath` is
called.