Add basic support for primality checking.
We'll probably want to refactor this somewhat so we can
configure if we want to do probablistic primality testing (default)
or certified testing.  We also probably want to memozie results;
I anticipate that a given development will use a small number of
concrete primes, but they may be mentioned many times.