Add settings for showing float in various formats.
This also changes the representation for Float used by the Concrete,
so that the float values carry around their sizes.  This is a bit redundant
as we have the same information in types, but this is not the case
in the `ppValue` function which, unfortunately, is not provided with a
type.