Add superclass information to the typechecker.
This teaches the typechecker about superclass information
so that, e.g., the constraint `Ring a` implies the constraint `Zero a`.
We do this by computing sets of constraints that are saturated under
the the superclass rules.  These are used during inference type to
compute the minimal collection of constraints necessary to imply
all the found goals, and at solving time to discharge goals that
are implied by constraints listed in type signatures.

Fixes #738