Refactor how prover configurations are chosen.
Rather than looking up prover information by name every
time we do a proof, we fetch the prover configuration as soon
as the user selects a prover using `:set prover=` and remember
it as part of the REPL state.  This will allow us to
do some work at configuration time and remember the results.

As part of this refactoring, we now print the list of solvers
that SBV found when using `:set prover=any`.