Fix the SBV `any` solver so that it waits for the first successful result.
Instead of relying on the internal SBV portfolio mechanism, we simply
do it ourselves "from the outside".  The SBV portfolio solver now works
in essentially the same way as the What4 one: if the first result we
get is an error, we continue waiting for other solvers and only
present exceptions to the user if all provers fail.

Fixes #798
Fixes #693