Change to markdown preprocessor.
We now allow indented fenced blocks.  Apparently those happen if you
want to put code in a list.  This being markdown, things are ambiguous,
but the main change is that if we see an indented line that starts with
``` now we treat it as starting a fenced block, rather than an
indented code block.

This fixes #805