Make `:m` work the same way as `:l`
More precisely, in both cases we unload all modules and then reload
everything (this is what `:l` used to do, while `:m` only reloaded
the current module).  This fixes #668.

There are opportunities to be smarter here: in particular when we
reload modules, we do need to parse them so that we can find out what
their dependencies are, and if needed to reload those.  However, if
none of the dependencies have changed, and we didn't change, then we
could reuse the current module.  This could be quite useful for modules
that take a long time to load.