Rearrange how `EvalOpts` are piped into the `trace` primitive.
Instead of making the `Eval` monad a reader that carries around
the EvalOpts, pass it directly into the primitives table, which
pipes it directly to the one place those options are needed.
Hopefully, this will reduce closure allocations in the evaluation
monad, and it also just generally simplifies things.

Note, we need to be careful to instantiate the `primTable` values
only as often as needed.  The What4 and Concrete primitive tables
now take arguments, which means they are not CAFs.  As a result,
they need to be let-bound someplace fairly global so that we don't
recompute the map every time we look up a primitive.  Hence, I have
removed the helper `evalPrim` functions, and forced call sites
instead to implement the map lookups.