Be more aggressive when checking signatures.
When checking a signature we try to solve sub-goals that mention any
of the variables mentioned in the signature's assumptions.  We do this
because these assumptions are local, and if the assumption is not solved
in the current context, we may not be able to solve it later.

This is OK if the goal does not contain free variables.  If it does,
things are a bit flaky as we may reject some programs and accept others
depending on in what order things happened to be instnatiated