Add source-location tracking when calling primitives.
This turns out to be surprisingly involved!  Every evaluation
helper and primitive that can raise a runtime error requires
knowledge of it's current source location, and we need to
rearrange slightly how variable lookups work so we can track
the source location of the calling context for primitives,
rather than just the location where the primitive was declared.