Rename the runtime-settable options to use a consistent style.
Options are now all named in camelCase style.  However, multiword
options may still be set using the old hypenated names as aliases,
so interaction scripts and the test suite do not have to be
changed. In addition, option lookup is altered to be case insensitive.

Thus, the canonical option name for the output SMT file when
an offline prover is selected is `smtFile`, which is what will
show up when running `:set` with no arguments, and what will be
suggested for tab-completion.  However, the option may be set
using the strings `smtfile` `SMTFile`, `smt-file`, `SMT-file`, etc.

Fixes #656