Add `FunDesc` information to the parsed AST.
This allows us to attach information to `EFun` nodes that describe
which named function the lambda belongs to, and which arguments
it handles.  This information is filled in by the `NoPat` pass
as it desugars multi-argument functions into a sequence of lambdas
and is used during typechecking to produce error messages.

There is an odd potential for the renamer to get things wrong here.
If a named function has an argument that shadows it's own name,
the renamer may incorrectly resolve the name in a function description
for later arguments.  The resulting error message will then point to
the wrong name (although it will be in the vicinity...).  It's hard
to fix this without folding the NoPat pass into the renamer (which
we should also do, but it's a nontrivial refactoring).