Proof-of-concept implenentation of imperative procedures.
Implement a new "procedure" mode that gives programmers
the ability to write imperative-style algorithms with
sequentual assignemnts, if/then/else, while and for loops.

This implementation is half-baked in a variety of ways,
but illustrates the ideas and lets us decide if this is worth
persuing.  The syntax, in particular, needs some work.

This implementation compiles procedures into recusive functional
programs inside the renamer.  As a result, the typechecker
and evaluator need no changes; the downside is that substandard
error messages are a very likely result.  A more realistic
implementation would probably typecheck procedures first,
then reduce them to functional programs; however, this would
have more downstream impacts as clients would need to be
aware of new typecheker AST nodes.